This was a bunch of changes that fixed (partially) compilation issues on some platforms.
It also allowed a static-compiled windows version, with lots of additions
for stupid compiler warnings (template name expansion being too large) which I
could not find a way to disable by default without including the
'warning_eliminator.h' file in nearly everything that uses templates, strings,
or those types.

+++ plugins/image_filter/lens_flare/flare_streaks.h
@@ -67,7 +67,7 @@ class TFlare_Streaks : public TFlare
         if ( ( _tCurStreak.fEndAngle >= (PI * 2.0f) ) &&
              (               _fAngle <  (PI / 2.0f) ) )
 	{
-	  _fAngle += 2.0f * PI;
+	  _fAngle += 2.0f * (float)PI;
 	}
 
         if ( ( _fAngle >= _tCurStreak.fStartAngle ) &&
